turning widgets into a module
